requirements.txt
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
ffmpeg-python==0.2.0
Flask==2.0.1
future==0.18.2
google-api-core==2.0.1
google-auth==2.1.0
google-cloud-core==2.0.0
google-cloud-speech==2.9.0
google-cloud-storage==1.42.2
google-crc32c==1.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.40.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
libcst==0.3.20
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.0
proto-plus==1.19.0
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dotenv==0.19.0
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.6
Werkzeug==2.0.1
